requirements.txt
pandas==1.3.5
scikit-learn==1.0.2
tensorflow==2.7.0
mlflow==1.22.0
feast[sqlite]==0.21.1
kfp==1.8.11
kubernetes==21.7.0
bentoml==1.0.0rc3
evidently==0.1.53.dev0
GitPython==3.1.27
pyyaml==6.0
requests==2.27.1